refactor: 스프링 부트 프로젝트로 변경
