+++ src/main/java/chess/chess/Board.java
@@ -1,5 +1,7 @@
 package chess.chess;
 
+import chess.chess.piece.*;
+
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
