+++ src/test/java/chess/chess/piece/PieceTest.java
@@ -3,6 +3,8 @@ package chess.chess.piece;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 public class PieceTest {
 
     @Test
